convert mã lỗi

+++ HeThongQuanLyNhaThuocLongChau/BusinessLogicLayer/SanPhamBLL.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HeThongQuanLyNhaThuocLongChau.Model
+namespace HeThongQuanLyNhaThuocLongChau.BusinessLogicLayer
 {
     public class SanPhamBLL
     {
@@ -14,47 +14,70 @@ namespace HeThongQuanLyNhaThuocLongChau.Model
 
         public DataTable findAll()
         {
-            string sql = "SELECT * FROM vv_SanPham";
-            return sanPhamDAL.findAll(sql);
+            return sanPhamDAL.findAll();
         }
 
-        public bool insert(string maSP, string tenSP, string donViTinh, string hanDung, double donGiaBan, string maLoai, string maNCC)
+        public bool insert(string maSP, string tenSP, string donViTinh, string hanDung, string donGiaBan, string maLoai, string maNCC)
         {
-            return sanPhamDAL.insert(maSP, tenSP, donViTinh, hanDung, donGiaBan, maLoai, maNCC);
+            if (string.IsNullOrEmpty(maSP) || string.IsNullOrEmpty(tenSP) || string.IsNullOrEmpty(donViTinh) || string.IsNullOrEmpty(hanDung)
+                || string.IsNullOrEmpty(donGiaBan) || string.IsNullOrEmpty(maLoai) || string.IsNullOrEmpty(maNCC))
+            {
+                return false;
+            }
+            else if (!double.TryParse(donGiaBan, out _) || double.Parse(donGiaBan) <= 0)
+            {
+                return false;
+            }
+
+            return sanPhamDAL.insert(maSP, tenSP, donViTinh, hanDung, double.Parse(donGiaBan), maLoai, maNCC);
         }
 
-        public bool update(string maSP, string tenSP, string donViTinh, string hanDung, double donGiaBan, string maLoai, string maNCC)
+        public bool update(string maSP, string tenSP, string donViTinh, string hanDung, string donGiaBan, string maLoai, string maNCC)
         {
-            return sanPhamDAL.update(maSP, tenSP, donViTinh, hanDung, donGiaBan, maLoai, maNCC);
+            if (string.IsNullOrEmpty(maSP) || string.IsNullOrEmpty(tenSP) || string.IsNullOrEmpty(donViTinh) || string.IsNullOrEmpty(hanDung)
+                || string.IsNullOrEmpty(donGiaBan) || string.IsNullOrEmpty(maLoai) || string.IsNullOrEmpty(maNCC))
+            {
+                return false;
+            }
+            else if (!double.TryParse(donGiaBan, out _) || double.Parse(donGiaBan) <= 0)
+            {
+                return false;
+            }
+
+            return sanPhamDAL.update(maSP, tenSP, donViTinh, hanDung, double.Parse(donGiaBan), maLoai, maNCC);
         }
 
         public bool deleteById(string maSP)
         {
+            if (string.IsNullOrEmpty(maSP))
+            {
+                return false;
+            }
             return sanPhamDAL.deleteById(maSP);
         }
 
         public DataTable search(string tenSP, string donViTinh, string hanDung, string donGiaBan, string tenLoai, string tenNCC)
         {
-            string sqlSearchPrice = "";
-            if (donGiaBan != "" && double.TryParse(donGiaBan, out _))
+            if (string.IsNullOrEmpty(tenSP) && string.IsNullOrEmpty(donViTinh) && string.IsNullOrEmpty(hanDung)
+                && string.IsNullOrEmpty(donGiaBan) && string.IsNullOrEmpty(tenLoai) && string.IsNullOrEmpty(tenNCC))
             {
-                sqlSearchPrice = "AND [Giá bán] = " + double.Parse(donGiaBan) + " ";
+                return null;
             }
-            String sql = "SELECT * FROM vv_SanPham " +
-                "WHERE [Tên SP] LIKE N'%" + tenSP + "%' " +
-                    "AND [Loại] LIKE N'%" + tenLoai + "%' " +
-                    "AND [Đơn vị tính] LIKE N'%" + donViTinh + "%' " +
-                    "AND [Hạn dùng] LIKE N'%" + hanDung + "%' " +
-                    "AND [Tên NCC] LIKE N'%" + tenNCC + "%' " +
-                    sqlSearchPrice;
 
-            return sanPhamDAL.findAll(sql);
+            if (!string.IsNullOrEmpty(donGiaBan))
+            {
+                if (!double.TryParse(donGiaBan, out _) || double.Parse(donGiaBan) <= 0)
+                {
+                    return null;
+                }
+            }
+
+            return sanPhamDAL.search(tenSP, donViTinh, hanDung, donGiaBan, tenLoai, tenNCC);
         }
 
         public bool checkExistsNameAndCategory(string tenSP, string tenLoai)
         {
-            string sql = "SELECT * FROM vv_SanPham";
-            DataTable dt = sanPhamDAL.findAll(sql);
+            DataTable dt = sanPhamDAL.findAll();
 
             foreach (DataRow dr in dt.Rows)
             {
